[dexdiff][tinker] Tidy up codes and prepare for optimizing IndexMap.
