+++ tinker-commons/src/main/java/com/tencent/tinker/commons/dexpatcher/algorithms/patch/FieldIdSectionPatchAlgorithm.java
@@ -20,8 +20,9 @@ import com.tencent.tinker.android.dex.Dex;
 import com.tencent.tinker.android.dex.FieldId;
 import com.tencent.tinker.android.dex.TableOfContents;
 import com.tencent.tinker.android.dex.io.DexDataBuffer;
-import com.tencent.tinker.android.dx.util.IndexMap;
 import com.tencent.tinker.commons.dexpatcher.struct.DexPatchFile;
+import com.tencent.tinker.commons.dexpatcher.util.AbstractIndexMap;
+import com.tencent.tinker.commons.dexpatcher.util.SparseIndexMap;
 
 /**
  * Created by tangyinsheng on 2016/7/4.
@@ -34,9 +35,9 @@ public class FieldIdSectionPatchAlgorithm extends DexSectionPatchAlgorithm<Field
             DexPatchFile patchFile,
             Dex oldDex,
             Dex patchedDex,
-            IndexMap oldToFullPatchedIndexMap
+            SparseIndexMap oldToPatchedIndexMap
     ) {
-        super(patchFile, oldDex, oldToFullPatchedIndexMap);
+        super(patchFile, oldDex, oldToPatchedIndexMap);
 
         if (patchedDex != null) {
             this.patchedFieldIdTocSec = patchedDex.getTableOfContents().fieldIds;
@@ -60,7 +61,7 @@ public class FieldIdSectionPatchAlgorithm extends DexSectionPatchAlgorithm<Field
     }
 
     @Override
-    protected FieldId adjustItem(IndexMap indexMap, FieldId item) {
+    protected FieldId adjustItem(AbstractIndexMap indexMap, FieldId item) {
         return indexMap.adjust(item);
     }
 
@@ -71,14 +72,14 @@ public class FieldIdSectionPatchAlgorithm extends DexSectionPatchAlgorithm<Field
     }
 
     @Override
-    protected void updateIndexOrOffset(IndexMap indexMap, int oldIndex, int oldOffset, int newIndex, int newOffset) {
+    protected void updateIndexOrOffset(SparseIndexMap sparseIndexMap, int oldIndex, int oldOffset, int newIndex, int newOffset) {
         if (oldIndex != newIndex) {
-            indexMap.mapFieldIds(oldIndex, newIndex);
+            sparseIndexMap.mapFieldIds(oldIndex, newIndex);
         }
     }
 
     @Override
-    protected void markDeletedIndexOrOffset(IndexMap indexMap, int deletedIndex, int deletedOffset) {
-        indexMap.markFieldIdDeleted(deletedIndex);
+    protected void markDeletedIndexOrOffset(SparseIndexMap sparseIndexMap, int deletedIndex, int deletedOffset) {
+        sparseIndexMap.markFieldIdDeleted(deletedIndex);
     }
 }
